refactor: break circular type dependencies
