+++ src/base/state/checkpoint-trust-port.ts
@@ -1,0 +1,3 @@
+export interface CheckpointTrustPort {
+  setOverride(domain: string, balance: number, reason: string): Promise<void>;
+}
