+++ src/orchestrator/execution/parallel-execution-types.ts
@@ -1,0 +1,6 @@
+export interface SubtaskResult {
+  task_id: string;
+  verdict: "pass" | "partial" | "fail";
+  output: string;
+  error?: string;
+}
