+++ src/reporting/reporting-engine.ts
@@ -13,32 +13,11 @@ import {
   buildExecutionSummaryContent,
   buildNotificationContent,
 } from "./report-formatters.js";
-
-// ─── Types ───
-
-export type ExecutionSummaryParams = {
-  goalId: string;
-  loopIndex: number;
-  observation: { dimensionName: string; progress: number; confidence: number }[];
-  gapAggregate: number;
-  taskResult: { taskId: string; action: string; dimension: string } | null;
-  stallDetected: boolean;
-  pivotOccurred: boolean;
-  elapsedMs: number;
-};
-
-export type NotificationType =
-  | "urgent"
-  | "approval_required"
-  | "stall_escalation"
-  | "completed"
-  | "capability_insufficient";
-
-export type NotificationContext = {
-  goalId: string;
-  message: string;
-  details?: string;
-};
+import type {
+  ExecutionSummaryParams,
+  NotificationContext,
+  NotificationType,
+} from "./reporting-types.js";
 
 // ─── ReportingEngine ───
 
@@ -546,4 +525,9 @@ export class ReportingEngine {
 }
 
 // ─── Re-exports ───
+export type {
+  ExecutionSummaryParams,
+  NotificationContext,
+  NotificationType,
+} from "./reporting-types.js";
 export { formatReportForCLI, buildExecutionSummaryContent, buildNotificationContent } from "./report-formatters.js";
